chore: welcome/page.tsx内のコンポーネントを別のディレクトリに移動
